\name{HPDGLM-package}
\alias{HPDGLM-package}
\alias{HPDGLM}
\docType{package}
\title{ HPDGLM - Distributed Regression for Big Data}
\description{
  \pkg{HPDGLM} provides distributed Generalized Linear Model. It is written based on the infrastructure created in HP-Lab for distributed computing in R.
}
\details{
 \tabular{ll}{
  Package: \tab HPDGLM\cr
  Type: \tab Package\cr
  Version: \tab 0.7.0\cr
  Date: \tab 2014-01-22\cr
 }
 Main Functions:
 \itemize{
  \item {hpdglm:} {It is a distributed version of glm.}
  \item {v.hpdglm:} {This function is implemented for evaluating a model built by hpdglm using Split-Sample-Validation method.}
  \item {cv.hpdglm:} {This function is implemented for evaluating a model built by hpdglm using Cross-Validation method.}
 }

}

\author{ Arash Fard <afard@vertica.com> }

\keyword{distributed R, distributed Regression, Big Data Analytics}


